# services/core-api/continuum_api_spec.yaml
openapi: 3.0.3
...
info:
  title: Continuum CRM API
  version: '1.0.0'
  description: |
    An AI-native CRM that captures customer interactions automatically and
    proposes the resulting record changes for human confirmation.

    ## Authentication

    All endpoints except signup, login, provider OAuth callbacks and provider
    webhooks require a bearer token:

        Authorization: Bearer <token>

    Two token types are accepted:

    - **Session JWT** — issued by `/auth/login`, used by the web application.
    - **API token** — prefixed `ctm_`, created in Settings. Subject to the rate
      limit below, and revocable with effect on the very next request.

    ## Rate limiting

    API tokens are limited to 300 requests per minute (FR-API-01). Every
    response carries `X-RateLimit-Limit`, `X-RateLimit-Remaining` and
    `X-RateLimit-Reset`. Requests below the limit are never throttled.

    ## Provenance

    Records that the agent can infer carry a `source` field
    (`HUMAN` | `AGENT_INFERRED` | `IMPORTED` | `ENRICHED`). Deals additionally
    carry `stageSource`, recording whether the current stage was set by a
    person or by a confirmed agent proposal. Consuming clients should surface
    this distinction rather than flattening it.

  contact:
    name: Continuum engineering
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.continuum.example/api/v1
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /deals/{id}/convert-to-project:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Deals, Projects]
      summary: Convert a won deal to a delivery project
      description: |
        One atomic action (FR-PROJ-01). The project inherits the company, every
        stakeholder contact, and the full activity history — the history by
        re-pointing existing links rather than copying, so capture continues
        onto the project with no re-linking step (FR-PROJ-03).

        Converting the same deal twice is rejected with a pointer to the
        existing project (FR-PROJ-06).
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                ownerId: { type: string, format: uuid, nullable: true }
      responses:
        '201':
          description: Converted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '409':
          description: Already converted, or the deal is not won
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /integrations/{provider}/connect:
    parameters:
      - name: provider
        in: path
        required: true
        schema: { type: string, enum: [google, microsoft] }
    post:
      tags: [Integrations]
      summary: Begin connecting a mailbox
      description: |
        One action starts both mail and calendar sync (FR-INT-01, FR-INT-02).
      responses:
        '200':
          description: Authorization URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  authorizationUrl: { type: string }
                  state: { type: string }
                  simulated: { type: boolean }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A session JWT, or a `ctm_`-prefixed API token.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Stable machine-readable code — branch on this, not on the message.
            message: { type: string }
            details: {}
...
    Project:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        status: { type: string, enum: [ACTIVE, ON_HOLD, COMPLETED, CANCELLED] }
        companyId: { type: string, format: uuid, nullable: true }
        originatingDealId: { type: string, format: uuid, nullable: true }
        milestones: { type: array, items: { type: object } }
...